refactor and style
